% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_variables.R
\name{arrange_variables}
\alias{arrange_variables}
\title{Arrange variables}
\usage{
arrange_variables(var_list)
}
\arguments{
\item{var_list}{A list of lists. Each second-level list corresponds to a
variable in the model and must be a name-equation pair. Given that variables
depend on other elements in the model, equations cannot consist of a single
scalar.}
}
\value{
A list consisting of the elements in \code{var_list} but arranged in
computational order.
}
\description{
\code{arrange_variables} returns a list of auxiliary variables sorted in
   computational order
}
\details{
This function iterates over each element until the variables are ordered in
computational order.
}
\examples{
unordered_vars <- list(list(name = "b", equation = "a + 1"),
                       list(name = "a", equation = "alpha"))
ordered_vars   <- arrange_variables(unordered_vars)
}
