% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_graph_funs.R
\name{call_graphical_funs}
\alias{call_graphical_funs}
\title{Call graphical functions}
\usage{
call_graphical_funs(mdl_obj)
}
\arguments{
\item{mdl_obj}{The output of \code{read_xmile}}
}
\description{
Puts graphical functions of a model in the global environment.
}
\examples{
\donttest{
mdl <- read_xmile("my_model.xmile")
call_graphical_funs(mdl)
}
}
