% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_else_builtins.R
\name{create_pt_statement}
\alias{create_pt_statement}
\title{Create Pulse Train statement}
\usage{
create_pt_statement(start_pt, duration_pt, repeat_pt, end_pt)
}
\arguments{
\item{start_pt}{Numeric}

\item{duration_pt}{Numeric}

\item{repeat_pt}{Numeric}

\item{end_pt}{Numeric}
}
\value{
A string
}
\description{
Create Pulse Train statement
}
\examples{
\dontrun{
create_pt_statement(5, 3, 10, 20)
}
}
