% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmile_helpers.R
\name{eval_constant_expr}
\alias{eval_constant_expr}
\title{Evaluate constant expression}
\usage{
eval_constant_expr(equation)
}
\arguments{
\item{equation}{A string}
}
\value{
A string
}
\description{
Evaluate constant expression
}
\examples{
\donttest{
eval_constant_expr(3 + 3) # returns 6
eval_constant_expr(a + 3) # returns a + 3
}
}
