% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_trans.R
\name{expit}
\alias{expit}
\title{Expit transformation}
\usage{
expit(x)
}
\arguments{
\item{x}{A real number}
}
\value{
A number in the range 0 to 1
}
\description{
Expit transformation
}
\examples{
expit(-3)
}
