% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variables.R
\name{extract_variables}
\alias{extract_variables}
\title{Extract variables from an equation}
\usage{
extract_variables(lhs, rhs)
}
\arguments{
\item{lhs}{A string with the name of the equation's subject, represented by
\code{y} in the canonical form.}

\item{rhs}{The equation or expression for the equation's subject, represented
by \code{f(x)} in the canonical form.}
}
\value{
A character vector with the unique elements in the \code{rhs}
}
\description{
\code{extract_variables} returns the unique elements in an equation.
}
\details{
This function assumes an expression in the canonical form \code{y = f(x)},
where x is a vector of n elements. \code{extract_variables} identify all
arguments in \code{f}.
}
\examples{
\dontrun{
lhs <- "z"
rhs <- "x ^ 2 + 2 * x * y + y ^ 2"
extract_variables(lhs, rhs)
}
}
