% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_trans.R
\name{logit}
\alias{logit}
\title{Logit transformation}
\usage{
logit(p)
}
\arguments{
\item{p}{A real number that represents a probability}
}
\value{
An unconstrained real number
}
\description{
Logit transformation
}
\examples{
logit(0.5)
}
