% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vensim.R
\name{parse_Vensim}
\alias{parse_Vensim}
\title{Parse Vensim}
\usage{
parse_Vensim(file)
}
\arguments{
\item{file}{A path to a Vensim file with extension .mdl}
}
\value{
A list...
 @export
}
\description{
Parse Vensim
}
