% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreters.R
\name{sd_conf_intervals}
\alias{sd_conf_intervals}
\title{Calculate confidence intervals}
\usage{
sd_conf_intervals(estimates, par_list, hsn, conf_level = 0.95)
}
\arguments{
\item{estimates}{A list or data frame}

\item{par_list}{A list}

\item{hsn}{Hessian matrix}

\item{conf_level}{A numeric input indicating the confidence level}
}
\value{
A data frame.
}
\description{
Calculate confidence intervals
}
\examples{
estimates <- c(-0.2630303, 1.5788579)
par_list  <- list(list(par_name  = "par_inv_R0",
                       par_trans = "expit"),
                  list(par_name  = "I0",
                       par_trans = "exp"))
hsn <- matrix(c(3513.10521, -493.5469626,
                -493.5469626, 88.4871290), ncol = 2)
sd_conf_intervals(estimates, par_list, hsn)
}
