% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{sd_constants}
\alias{sd_constants}
\title{Summarise the information of the model's constants in a data frame}
\usage{
sd_constants(mdl)
}
\arguments{
\item{mdl}{A list which is the output from read_xmile.}
}
\value{
A data frame.
}
\description{
Summarise the information of the model's constants in a data frame
}
\examples{
path <- system.file("models", "SIR.stmx", package = "readsdr")
mdl  <- read_xmile(path)
sd_constants(mdl)
}
