% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.R
\name{sd_fixed_delay}
\alias{sd_fixed_delay}
\title{Fixed delay}
\usage{
sd_fixed_delay(var, time, delay, init, .memory)
}
\arguments{
\item{var}{A string that indicates the delayed variable.}

\item{time}{A number that indicates current simulation time.}

\item{delay}{A number that indicates the delay time.}

\item{init}{A number that indicates the function's output value of at the
start of the simulation.}

\item{.memory}{A data frame that keeps past values of delayed variables.}
}
\value{
A number.
}
\description{
Fixed delay
}
\examples{
.memory <- data.frame(time = 3, inflow = 3)
rownames(.memory) <- 3
sd_fixed_delay("inflow", 5, 2, 0, .memory)
}
