% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_fun.R
\name{sd_loglik_fun}
\alias{sd_loglik_fun}
\title{Generate log-likelihood function for an SD model}
\usage{
sd_loglik_fun(
  pars_df,
  deSolve_components,
  sim_controls,
  fit_options,
  extra_stocks = NULL
)
}
\arguments{
\item{pars_df}{A data frame}

\item{deSolve_components}{A list}

\item{sim_controls}{A list}

\item{fit_options}{A list}

\item{extra_stocks}{An optional list}
}
\value{
A function
}
\description{
Generate log-likelihood function for an SD model
}
