% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_measurements.R
\name{sd_measurements}
\alias{sd_measurements}
\title{Generate measurements}
\usage{
sd_measurements(
  n_meas,
  meas_model,
  ds_inputs,
  start_time = NULL,
  stop_time = NULL,
  timestep = NULL,
  integ_method = "euler"
)
}
\arguments{
\item{n_meas}{Number of measurements}

\item{meas_model}{Measurement model. A list of sampling distributions. Each
sampling distribution is a list of three elements:

  - \code{stock_name} (a string): refers to the stock upon which measurements are made;

  - \code{meas_type} (a string): indicates whether the measurement is performed on the
 value of the stock ('as-is') or its net change ('net_change');

  - \code{dist} (a list): The probability distribution used to model measurements on
  the stock. This list also consists of three elements:
  name (RNG function such as rnorm), stock_par (indicates which distribution's
  parameter corresponds to the measured stock), and extra_pars, should the
  probability required additional parameters.}

\item{ds_inputs}{A list of deSolve inputs generated by read_xmile}

\item{start_time}{A number}

\item{stop_time}{A number}

\item{timestep}{A number}

\item{integ_method}{A string. Either "euler" or "rk4"}
}
\value{
A data frame.
}
\description{
Generate measurements
}
\examples{
  filepath <- system.file("models/", "SEIR.stmx", package = "readsdr")
  mdl      <- read_xmile(filepath)

  mm1 <- list(stock_name = "C",
              meas_type  = "as_is",
              dist       = list(name       = "rnbinom",
                                stock_par  = "mu",
                                extra_pars = list(size = 0.1)))
  meas_model <- list(mm1)

  sd_measurements(n_meas       = 2,
                  meas_model   = meas_model,
                  ds_inputs    = mdl$deSolve_components,
                  start_time   = 0,
                  stop_time    = 10,
                  timestep     = 1/16,
                  integ_method = "rk4")
}
