% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{sd_prior}
\alias{sd_prior}
\title{SD prior}
\usage{
sd_prior(par_name, dist, dist_pars, type = "constant")
}
\arguments{
\item{par_name}{A string}

\item{dist}{A string}

\item{dist_pars}{A vector}

\item{type}{A string. It can be either 'constant' or 'init'. It is 'constant'
by default.}
}
\value{
A list
}
\description{
SD prior
}
\examples{
sd_prior("par_beta", "lognormal", c(0, 1))
}
