% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_else_builtins.R
\name{sd_pulse_s}
\alias{sd_pulse_s}
\title{Replicate the behaviour of pulse from Stella}
\usage{
sd_pulse_s(time, volume, start_p, interval)
}
\arguments{
\item{time}{A number}

\item{volume}{A number}

\item{start_p}{A number}

\item{interval}{A number}
}
\value{
A number
}
\description{
Replicate the behaviour of pulse from Stella
}
