% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_else_builtins.R
\name{sd_pulse_train}
\alias{sd_pulse_train}
\title{PULSE TRAIN}
\usage{
sd_pulse_train(time, start_pulse, duration_pulse, repeat_pt, end_pulse)
}
\arguments{
\item{time}{A numeric argument that indicates the current simulation time}

\item{start_pulse}{A numeric argument that indicates the start of the pulse}

\item{duration_pulse}{A numeric argument that indicates the width of the pulse}

\item{repeat_pt}{A numeric argument that indicates the repetition pattern}

\item{end_pulse}{A numeric argument that indicates the end of the sequence}
}
\value{
1 during the pulse, 0 otherwise.
}
\description{
PULSE TRAIN
}
\examples{
sd_pulse_train(5, 5, 3, 10, 20)
}
