% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_else_builtins.R
\name{sd_pulse_v}
\alias{sd_pulse_v}
\title{Replicate the behaviour of the PULSE function from Vensim}
\usage{
sd_pulse_v(time, startPulse, duration)
}
\arguments{
\item{time}{A number}

\item{startPulse}{A number}

\item{duration}{A number}
}
\value{
A number
}
\description{
Replicate the behaviour of the PULSE function from Vensim
}
\examples{
timestep <- function() 0.25 # replicates timestep() from deSolve
sd_pulse_v(1, 1, 2)
}
