% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocities.R
\name{sd_velocities}
\alias{sd_velocities}
\title{Construct velocity equations in terms of stocks and constants}
\usage{
sd_velocities(desc_list)
}
\arguments{
\item{desc_list}{Element 'description' from the list returned by \code{read_xmile()}}
}
\value{
A data frame of two columns.
}
\description{
Construct velocity equations in terms of stocks and constants
}
\examples{
 filepath  <- system.file("models/", "SIR.stmx", package = "readsdr")
 mdl       <- read_xmile(filepath)
 desc_list <- mdl$description
 sd_velocities(desc_list)
}
