% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{stan_data}
\alias{stan_data}
\title{Stan's data block for ODE models}
\usage{
stan_data(vars_vector)
}
\arguments{
\item{vars_vector}{a string vector. Each element corresponds to a vector's
name for which users will supply data.}
}
\value{
a string that contains the Stan code for the data block.
}
\description{
Stan's data block for ODE models
}
\examples{
stan_data("y")
}
