% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{stan_transformed_data}
\alias{stan_transformed_data}
\title{Stan's transformed data block for ODE models}
\usage{
stan_transformed_data()
}
\value{
a string
}
\description{
Stan's transformed data block for ODE models
}
\examples{
td <- stan_transformed_data()
}
