% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{sd_prior}
\alias{sd_prior}
\title{SD prior}
\usage{
sd_prior(par_name, dist, dist_pars, type = "constant", min_0 = FALSE)
}
\arguments{
\item{par_name}{A string}

\item{dist}{A string}

\item{dist_pars}{A vector}

\item{type}{A string. It can be either 'constant' or 'init'. It is 'constant'
by default. 'init' refers to parameters that have only affect the model at
time 0.}

\item{min_0}{An optional boolean indicating whether the prior has a lower
bound at zero. In the current implementation, this parameter only has an
effect on normal priors.}
}
\value{
A list
}
\description{
SD prior
}
\examples{
sd_prior("par_beta", "lognormal", c(0, 1))
sd_prior("par_rho", "normal", c(0, 1), min_0 = TRUE)
}
