% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_else_builtins.R
\name{sd_pulse_s}
\alias{sd_pulse_s}
\title{Replicate the behaviour of the PULSE function from Stella}
\usage{
sd_pulse_s(time, volume, start_p, interval)
}
\arguments{
\item{time}{A number}

\item{volume}{A number}

\item{start_p}{A number}

\item{interval}{A number}
}
\value{
A number
}
\description{
This function must be placed inside the object that will be passed as the
argument \code{func} to deSolve's \code{ode} function.
}
\examples{
timestep <- function() 0.25 # replicates timestep() from deSolve
sd_pulse_s(2, 1, 2, 0)

}
